Make build commands PHONY
For any Makefile target that does not actually produce a file of
that name, the target should be declared PHONY. This prevents a
file or directory with the same name from being interpreted as the
target of the command.

For example, the `build` command in this Makefile produces no
result because there is already a `build` directory, and Make
interprets that as "being up to date".